FE-876: Moved file
