+++ src/pages/blacklist/components/tests/IncidentsCollection.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import IncidentsCollection from '../IncidentsCollection';
-import TestApp from 'src/helpers/TestApp';
+import TestApp from 'src/__testHelpers__/TestApp';
 
 describe('Blacklist Component: IncidentsCollection', () => {
   const now = new Date('2019-12-18T04:20:00-04:00');
